requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
chromadb==0.4.18
langchain==0.1.0
langchain-google-genai==0.0.6
tiktoken==0.5.2
numpy==1.26.2
pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
